extract buttons from radio/index
